utils calibration tests
