--- src/data-mgt/java/streams/device-measurements/src/main/java/net/airqo/models/CalibratedValue.java
@@ -1,4 +1,0 @@
-package net.airqo.models;
-
-public class CalibratedValue {
-}
